api fixes on banners api, listing and detail

+++ database/migrations/2025_11_22_202000_add_status_to_users_table.php
@@ -12,7 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->tinyInteger('status')->default(1)->after('password');
+            if(!Schema::hasColumn('users','status')){
+                $table->tinyInteger('status')->default(1)->after('password');
+            }
         });
     }
 
@@ -22,7 +24,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('status');
+              if(!Schema::hasColumn('users','status')){
+                  $table->dropColumn('status');
+              }
         });
     }
 };
